refactor(tui): separate renderer by page
